mobile-backend/requirements.txt
Flask-Cors
Flask==3.0.0
python-dotenv==1.0.1
google-generativeai>=0.5.0
